Add New Build Flavor, Regtest
Add a capability to the app to connect to a node that runs regtest version of Bitcoin. To do that I added a build flaver regtest to the wallet/build.gradle.kts. Based on this flaver the NETWORK_PARAMETERS which is checked through the app in different places will be given the value RegTestParams, so that You can make conditional changes in the code. the flaver also enabled us to add resouces, like different colors, based on which build has been chosen. For now we chosed the dark green and light blue for the regtest theme.